Use appropriate Library Manager category in metadata
The `category` field of the library.properties metadata file allows specifying a general category for the library, which
can be used to sort Library Manager listings.

The previous category value "Display" was not really appropriate for this library. Lacking a category that does fit the
nature of this library, we are left with "Other" as the best choice.